Progressed In the Project
Updated the UI of Quest and DashBoard
Attached UI to DataAccessLayer thorugh ViewModel

+++ DataModelLayer/DataModels/PlayerStats.cs
@@ -8,26 +8,26 @@ namespace DataModelLayer.DataModels
 
         public string? Id { get; set; }
 
-        public int HumanPlayerLevel { get; set; }
+        public long HumanPlayerLevel { get; set; }
 
         public long HumanPlayerExp { get; set; }
 
         public long HumanPlayerLevelMaxExp { get; set; }
 
-        public int STR { get; set; }
+        public double STR { get; set; }
 
-        public int AGI { get; set; }
+        public double AGI { get; set; }
 
-        public int INT { get; set; }
+        public double INT { get; set; }
 
-        public int COMM { get; set; }
+        public double COMM { get; set; }
 
-        public int MENSTB { get; set; }
+        public double MENSTB { get; set; }
 
-        public int CRT { get; set; }
+        public double CRT { get; set; }
 
-        public int HP { get; set; }
+        public double HP { get; set; }
 
-        public int HYG { get; set; }
+        public double HYG { get; set; }
     }
 }
